+++ htdocs/contact.php
@@ -242,7 +242,7 @@ function validateEmail(email) {
 </script>
 
 <?php
-    que_modal_tpl('login');
+    que_modal_tpl('login','logout');
     echo write_modal_tpls();
 ?>