+++ htdocs/no-reciepe.php
@@ -122,7 +122,7 @@ output_session_notification();
 ?>
 
 <?php
-    que_modal_tpl('login');
+    que_modal_tpl('login','logout');
     echo write_modal_tpls();
 ?>